requirements.txt
apturl==0.5.2
asgiref==3.2.3
asn1crypto==0.24.0
autopep8==1.4.4
Brlapi==0.6.6
cachetools==3.1.1
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==3.0
evdev==0.7.0
google-api-core==1.14.3
google-auth==1.8.2
google-cloud-core==1.1.0
google-cloud-translate==1.3.1
googleapis-common-protos==1.6.0
grpcio==1.25.0
httplib2==0.9.2
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.6.1
python-debian==0.1.32
python-xlib==0.20
pytz==2019.3
pyudev==0.21.0
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
rsa==4.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
xkit==0.0.0
zope.interface==4.3.2
